refactor: unify platform URL resolution with config-file-first fallback chain
